fixed dll

+++ Project_1/Testing/Testing.hpp
@@ -9,6 +9,6 @@ class Testing{
 
     double test_time(DataStructure& object, std::function<void()> func);
     
-    void save_to_csv(std::string saveFilePath, std::string key_name, std::string value_name, int keys[], float values[], int values_count);
+    void save_to_csv(std::string saveFilePath, std::string key_name, std::string value_name, int keys[], double values[], int values_count);
 
 };
